V12 - corrections et suppressions

+++ backend/controllers/post.js
@@ -3,15 +3,9 @@ const { models } = require("../sequelize/sequelize");
 const sequelize = require("../sequelize/sequelize");
 //gestion des fichier
 const fs = require("fs");
+
+//*************** Articles (Post) *********************
 //post
-// cree une sauce
-
-// getProfilById = async (profilId) => {
-//   models.profil.findOne({ where: { id: profilId } }).then((profil) => {
-//     return profil.dataValues;
-//   });
-// };
-
 exports.createPost = (req, res, next) => {
   let imageUrl = "";
   if (req.file)
@@ -30,63 +24,25 @@ exports.createPost = (req, res, next) => {
     .catch((error) => res.status(400).json({ error }));
 };
 
-//met a jour le status j'aime de l'utilisateur sur un Post
-exports.updateLikePost = async (req, res, next) => {
-  models.liked
-    .findOne({
-      where: { articleId: req.params.id, profilId: req.body.profilId },
-    })
-    .then(async (likeTrouver) => {
-      let strLikeReturn = "";
-      let idarticle = req.params.id;
-      if (likeTrouver) {
-        if (req.body.like == 0) {
-          //detriut
-          likeTrouver.destroy();
-          models.article.decrement("like", {
-            by: 1,
-            silent: true,
-            where: { id: req.params.id },
-          });
-        }
-      } else {
-        if (req.body.like == 1) {
-          //cree
-          models.liked.create({
-            profilId: req.body.profilId,
-            articleId: req.params.id,
-          });
-          models.article.increment("like", {
-            by: 1,
-            silent: true,
-            where: { id: req.params.id },
-          });
-        }
-      }
-      res.status(200).json({ message: "Like modifié !" });
-    })
-    .catch((error) => res.status(400).json({ error }));
-};
-
-//put
 // met à jour d'un post
 exports.ModifyPost = async (req, res, next) => {
   let imageUrl = "";
-
   // supprime l'ancienne image
   await models.article
     .findOne({ where: { id: req.params.id } })
     .then((post) => {
-      if (req.file) {
-        if (post.img && post.img != "") {
-          const filename = post.img.split("/images/")[1];
-          fs.unlinkSync(`images/${filename}`); //suppression synchrone
+      if (post) {
+        if (req.file) {
+          if (post.img && post.img != "") {
+            const filename = post.img.split("/images/")[1];
+            fs.unlinkSync(`images/${filename}`); //suppression synchrone
+          }
+          imageUrl = `${req.protocol}://${req.get("host")}/images/${
+            req.file.filename
+          }`;
+        } else {
+          if (req.body.img != "") imageUrl = post.img;
         }
-        imageUrl = `${req.protocol}://${req.get("host")}/images/${
-          req.file.filename
-        }`;
-      } else {
-        if (req.body.img != "") imageUrl = post.img;
       }
       return post;
     });
@@ -128,16 +84,12 @@ exports.deletePost = async (req, res, next) => {
 //get
 //renvoie tout les articles
 exports.getAllPost = (req, res, next) => {
-  //tableau des articles +profil
-
   models.article
-    .findAll()
+    .findAll({ order: [["createdAt", "DESC"]] })
     .then(async (articles) => {
       if (articles) {
         let listArticles = [];
-        // let userLike = {};
         for (const articleTemp of articles) {
-          // userLike = { likestate: 0 };
           let profilArticle = await models.profil
             .findOne({
               attributes: ["id", "pseudonyme", "avatar", "fonction"],
@@ -164,40 +116,14 @@ exports.getAllPost = (req, res, next) => {
       res.status(400).json({ error });
     });
 };
-//renvoie l'article avec l'id
-exports.getOnePost = (req, res, next) => {
-  models.article
-    .findOne({
-      where: { id: articleTemp.profilId },
-    })
-    .then((article) => {
-      models.profil.findOne(
-        {
-          attributes: ["id", "pseudonyme", "avatar", "fonction"],
-          where: { id: article.profilId },
-        }
-          .then((profil) => {
-            models.like
-              .findOne({
-                attributes: ["id", "profilId", "articleId"],
-                where: { articleId: articleTemp.id, profilId: profil.id },
-              })
-              .then((like) => {
-                res
-                  .status(200)
-                  .json({ article: article, profil: profil, like: like });
-              });
-          })
-          .catch((error) => res.status(404).json({ error }))
-      );
-    })
-    .catch((error) => res.status(404).json({ error }));
-};
+
+// ****************** Messages **********************
 //renvoie la liste des messages pour un article
 exports.getAllMessagesForOnePost = (req, res, next) => {
   models.message
     .findAll({
       where: { articleId: req.params.id },
+      order: [["createdAt", "DESC"]],
     })
     .then(async (messages) => {
       let listMessages = [];
@@ -261,6 +187,7 @@ exports.deleteMessageForOnePost = (req, res, next) => {
     .catch((error) => res.status(400).json({ error }));
 };
 
+// ************** Like *****************
 exports.getLikeState = (req, res, next) => {
   models.liked
     .findOne({
@@ -273,5 +200,89 @@ exports.getLikeState = (req, res, next) => {
       }
       res.status(200).json({ like: resultlike });
     })
+    .catch((error) => res.status(400).json({ error }));
+};
+//met a jour le status j'aime de l'utilisateur sur un Post
+exports.updateLikePost = async (req, res, next) => {
+  models.liked
+    .findOne({
+      where: { articleId: req.params.id, profilId: req.body.profilId },
+    })
+    .then(async (likeTrouver) => {
+      let strLikeReturn = "";
+      let idarticle = req.params.id;
+      if (likeTrouver) {
+        if (req.body.like == 0) {
+          //detriut
+          likeTrouver.destroy();
+          models.article.decrement("like", {
+            by: 1,
+            silent: true,
+            where: { id: req.params.id },
+          });
+        }
+      } else {
+        if (req.body.like == 1) {
+          //cree
+          models.liked.create({
+            profilId: req.body.profilId,
+            articleId: req.params.id,
+          });
+          models.article.increment("like", {
+            by: 1,
+            silent: true,
+            where: { id: req.params.id },
+          });
+        }
+      }
+      res.status(200).json({ message: "Like modifié !" });
+    })
+    .catch((error) => res.status(400).json({ error }));
+};
+//*********** Profil **************************/
+exports.updateProfil = async (req, res, next) => {
+  let imageUrl = "";
+  // supprime l'ancienne image
+  await models.profil
+    .findOne({ where: { id: req.body.profilId } })
+    .then((profil) => {
+      if (req.file) {
+        if (profil.avatar && profil.avatar != "") {
+          const filename = profil.avatar.split("/images/")[1];
+          fs.unlinkSync(`images/${filename}`); //suppression synchrone
+        }
+        imageUrl = `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`;
+      } else {
+        if (req.body.avatar != "") {
+          imageUrl = profil.avatar;
+        }
+      }
+      return profil;
+    });
+
+  let psedoTemp = req.body.pseudonyme ? req.body.pseudonyme : "";
+  let fonctionTemp = req.body.fonction ? req.body.fonction : "";
+
+  models.profil
+    .update(
+      {
+        pseudonyme: psedoTemp,
+        fonction: fonctionTemp,
+        avatar: imageUrl,
+      },
+      {
+        where: { id: req.body.profilId },
+      }
+    )
+    .then(() =>
+      res.status(200).json({
+        message: "Profil modifié !",
+        avatar: imageUrl,
+        pseudonyme: req.body.pseudonyme,
+        fonction: req.body.fonction,
+      })
+    )
     .catch((error) => res.status(400).json({ error }));
 };
